Removed OperationsTree, code moved to GUIMode.tsx
