+++ client/src/components/Visualiser/Controller/GUIMode/GUIMode.tsx
@@ -1,7 +1,9 @@
-import { Box } from '@mui/material';
-import React, { FC } from 'react';
-import { DataStructure, Documentation } from 'visualiser-src/common/typedefs';
-import OperationsTree from './GUIOperations/OperationsTree';
+import React, { FC, useContext } from 'react';
+// import OperationsTree from './GUIOperations/OperationsTree';
+import { Alert, Box, List, Typography } from '@mui/material';
+import VisualiserContext from 'components/Visualiser/VisualiserContext';
+import OperationDetails from './OperationDetails';
+
 /**
  * The GUI form that lets users input arguments to a menu of commands and then
  * have them affect the visualiser.
@@ -12,10 +14,30 @@ import OperationsTree from './GUIOperations/OperationsTree';
  * All the commands that the form supports are listed separately in the
  * `commandsInputRules.ts` file, where the terminal commands also reside.
  */
-const GUIMode = () => (
-  <Box sx={{ height: 'calc(100% - 64px)', overflow: 'auto' }}>
-    <OperationsTree />
-  </Box>
-);
+const GUIMode = () => {
+  const { documentation, topicTitle } = useContext(VisualiserContext);
+
+  return !documentation ? (
+    <Alert severity="error">
+      No operations are defined for the topicTitle &apos;
+      {topicTitle}
+      &apos;
+    </Alert>
+  ) : (
+    <Box sx={{ padding: 2, overflow: 'auto', height: 'calc(100% - 64px)' }}>
+      <Typography color="textPrimary">{topicTitle}</Typography>
+      <List>
+        {Object.keys(documentation).map((command, idx) => (
+          <Box key={documentation[command].id}>
+            <OperationDetails
+              command={command}
+              isLast={idx === Object.keys(documentation).length - 1}
+            />
+          </Box>
+        ))}
+      </List>
+    </Box>
+  );
+};
 
 export default GUIMode;
