+++ client/src/components/Visualiser/Controller/GUIMode/OperationDetails.tsx
